# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
...
  handlers.EnvelopeError:
    properties:
      error:
        example: invalid id
        type: string
      success:
        example: false
        type: boolean
    type: object
  handlers.EnvelopeOK:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  handlers.accountView:
    properties:
      email:
        type: string
      github_connected:
        type: boolean
      google_connected:
        type: boolean
      has_password:
        type: boolean
      id:
        type: integer
      is_admin:
        type: boolean
      name:
        type: string
    type: object
...
  handlers.bootstrapRequest:
    properties:
      force:
        type: boolean
      nodes:
        items:
          type: string
        type: array
      os:
        items:
          type: string
        type: array
    type: object
  handlers.bootstrapStatusView:
    properties:
      bootstrapped:
        type: boolean
    type: object
...
  handlers.changePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  handlers.verifyAccessCodeRequest:
    properties:
      code:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Self-hosted VM provisioning portal for Proxmox VE. The API is
    wrapped in a uniform `{success, data, error}` envelope on every
    response — see EnvelopeOK / EnvelopeError schemas below.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nimbus HTTP API
  version: "1.0"
paths:
  /access-code/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnvelopeOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Whether the user has verified the current access code
      tags:
      - auth
  /access-code/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Access code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.verifyAccessCodeRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnvelopeOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Verify the current access code
      tags:
      - auth
  /account:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/handlers.accountView'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Current user's profile + linked-provider state
      tags:
      - auth
  /account/password:
    put:
      consumes:
      - application/json
      description: |-
        Requires the current password — a stolen session alone
        shouldn't let an attacker lock the real owner out.
      parameters:
      - description: Current + new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.changePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EnvelopeOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Rotate the current user's password
      tags:
      - auth
  /admin/bootstrap-status:
    get:
      description: |-
        Read-only yes/no — both admins and members need it so the
        Provision UI can decide whether to render the form (templates
        ready) or the "admin access required" card.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/handlers.bootstrapStatusView'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Whether at least one VM template exists
      tags:
      - bootstrap
  /admin/bootstrap-templates:
    post:
      consumes:
      - application/json
      description: |-
        Long-running (up to ~20 min) — downloads cloud images, clones,
        and converts to templates. Empty body uses defaults (all
        catalogue OSes on all online nodes). The 30-minute route
        timeout in the router accommodates the slowest path.
      parameters:
      - description: Optional override of node/OS scope and force-rebuild flag
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.bootstrapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/bootstrap.Result'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: Build VM templates on cluster nodes (admin)
      tags:
      - bootstrap
  /audit:
    get:
      description: |-
        Read-only inspection of the cluster audit log. Newest
        first; filterable by actor, action prefix, and time
        range. Use the prefix form to scope by domain — e.g.
        `action_prefix=vm.` matches every VM event;
        `action_prefix=settings.` matches all settings updates.
      parameters:
      - description: filter by actor user id
        in: query
        name: actor_id
        type: integer
      - description: filter by action prefix (e.g. `vm.`, `settings.`)
        in: query
        name: action_prefix
        type: string
      - description: RFC3339 lower bound on created_at
        in: query
        name: since
        type: string
      - description: RFC3339 upper bound on created_at
        in: query
        name: until
        type: string
      - description: page size (1-500, default 100)
        in: query
        name: limit
        type: integer
      - description: row offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/handlers.auditListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      security:
      - cookieAuth: []
      summary: List audit events (admin)
      tags:
      - audit
  /auth/github:
    get:
      description: |-
        Sets short-lived state + intent cookies, then redirects to
        GitHub's authorization URL.
      responses:
        "307":
          description: Redirect to GitHub OAuth consent
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Start the GitHub OAuth sign-in flow (public)
      tags:
      - auth
  /auth/github/callback:
    get:
      description: |-
        Validates the state cookie, exchanges the code, and either
        signs the user in or links the GitHub account to the
        existing session (per the intent cookie). Always responds
        with a redirect — to the SPA on success, or back to /login
        with an explanatory query param on failure.
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state echoed by GitHub
        in: query
        name: state
        required: true
        type: string
      responses:
        "307":
          description: Redirect
      summary: GitHub OAuth callback (public)
      tags:
      - auth
  /auth/github/link:
    get:
      description: |-
        Same callback as sign-in mode but the intent cookie tells
        the callback to attach to the current session rather than
        mint a new one.
      responses:
        "307":
          description: Redirect to GitHub OAuth consent
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Start the GitHub OAuth flow in link mode
      tags:
      - auth
  /auth/google:
    get:
      responses:
        "307":
          description: Redirect to Google OAuth consent
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Start the Google OAuth sign-in flow (public)
      tags:
      - auth
  /auth/google/callback:
    get:
      description: See the GitHub callback for the shared sign-in/link semantics.
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state echoed by Google
        in: query
        name: state
        required: true
        type: string
      responses:
        "307":
          description: Redirect
      summary: Google OAuth callback (public)
      tags:
      - auth
  /auth/google/link:
    get:
      responses:
        "307":
          description: Redirect to Google OAuth consent
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Start the Google OAuth flow in link mode
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Sets the session cookie inline on success. 403 with
        message "Account suspended" when the user row is suspended.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/service.UserView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Email/password sign-in (public)
      tags:
      - auth
...
  /auth/magic/{token}:
    get:
      description: |-
        Always returns 307 — to /account?from=magic on success or to
        /login?magic=<reason> when the token is invalid/expired/used.
      parameters:
      - description: Magic-link token
        in: path
        name: token
        required: true
        type: string
      responses:
        "307":
          description: Redirect
      summary: Magic-link sign-in (public, single-use)
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Open-self-signup is gated by RequirePasswordlessAuth in
        production deployments — this endpoint always responds, but
        new accounts may be unable to sign in until they link OAuth.
      parameters:
      - description: New account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.EnvelopeOK'
            - properties:
                data:
                  $ref: '#/definitions/service.UserView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: Register a new user account (public)
      tags:
      - auth
...
  /gpu/install.sh:
    get:
      description: |-
        Validates ?token= against the active pairing window. The
        script body posts to /api/gpu/register and runs the
        install-inference / install-worker scripts.
      parameters:
      - description: Pairing token
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: Bash script body
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.EnvelopeError'
      summary: GX10 install bash script (public, pairing-token-auth)
      tags:
      - gpu
...
securityDefinitions:
  cookieAuth:
    description: |-
      Session cookie set by /api/auth/login. Required by all routes
      outside the public auth + setup surface.
    in: cookie
    name: nimbus_sid
    type: apiKey
  workerAuth:
    description: |-
      Bearer token used by the GX10 GPU worker daemon. Format:
      `Bearer <hex>`. Mint via the Settings → GPU page.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"